app/requirements.txt
fastapi==0.104.0
fastapi-users[sqlalchemy]==12.1.2
asyncpg
alembic==1.12.0
psycopg2-binary==2.9.9
pillow==10.1.0
pydantic-async-validation==0.1.0
requests==2.31.0
fastapi-mail[httpx]==1.4.1
